# backend/tests/fixtures/openapi_before_mvc.json
{
  "components": {
    "schemas": {
...
      "ExecuteRequest": {
        "properties": {
          "payment_id": {
            "title": "Payment Id",
            "type": "string"
          }
        },
        "required": [
          "payment_id"
        ],
        "title": "ExecuteRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TopUpRequest": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          }
        },
        "required": [
          "amount"
        ],
        "title": "TopUpRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WithdrawRequest": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "wallet_number": {
            "default": "",
            "title": "Wallet Number",
            "type": "string"
          }
        },
        "required": [
          "amount"
        ],
        "title": "WithdrawRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Student-to-student ride-sharing network — Backend API",
    "title": "Arooohi API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/wallet/payment/{payment_id}": {
      "get": {
        "description": "Reconciliation: ask the gateway what really happened to an interrupted payment.",
        "operationId": "payment_status_api_wallet_payment__payment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "payment_id",
            "required": true,
            "schema": {
              "title": "Payment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Status",
        "tags": [
          "Wallet & bKash"
        ]
      }
    },
...
    "/api/wallet/topup": {
      "post": {
        "description": "Step 1+2: grant token, then create the payment and hand back a checkout URL.",
        "operationId": "start_topup_api_wallet_topup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Topup",
        "tags": [
          "Wallet & bKash"
        ]
      }
    },
    "/api/wallet/topup/execute": {
      "post": {
        "description": "Step 4: the ONLY place a top-up can be credited.\n\nIdempotent twice over: the gateway returns the same trxID on re-execute, and\nuq_transactions_payment stops a second ledger row for the same paymentID.",
        "operationId": "execute_topup_api_wallet_topup_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Topup",
        "tags": [
          "Wallet & bKash"
        ]
      }
    },
    "/api/wallet/transactions": {
      "get": {
        "description": "Full ledger history — also the data source for receipts (#10).",
        "operationId": "list_transactions_api_wallet_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Transactions",
        "tags": [
          "Wallet & bKash"
        ]
      }
    },
    "/api/wallet/withdraw": {
      "post": {
        "description": "Cash out to bKash — the other edge of the prepaid model.",
        "operationId": "withdraw_api_wallet_withdraw_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Withdraw",
        "tags": [
          "Wallet & bKash"
        ]
      }
    },
...
  }
}